% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_coefficients.R
\name{Get_coefficients}
\alias{Get_coefficients}
\title{Get coefficients}
\usage{
Get_coefficients(Family, Lmax.mean)
}
\arguments{
\item{Family}{Taxaon Name}

\item{Lmax.mean}{Mean of 'largest locally measured length'.}
}
\description{
Retrieves Life History values for Family
}
